MYPROJECT/ubuntu-cap-backend/requirements.txt
africastalking==2.0.1
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-filter==25.2
djangorestframework==3.16.1
idna==3.11
joblib==1.5.2
numpy==2.3.4
pandas==2.3.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
responses==0.25.8
schema==0.7.8
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.5.0
xgboost==3.1.1
